refactor(note): decompose PatientDetails into focused hooks
